# api/openapi/creative-v1.yaml
openapi: 3.1.0
info:
  title: cookies Creative API
  version: 0.1.0
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
  description: >-
    The Creative vertical owns CreativeIntakes, CreativeTasks and content
    drafts. Organization scope comes from trusted identity and project scope is
    re-authorized from the path.
servers:
  - url: /
security:
  - cookiesAuth: []
paths:
...
  /api/creative/v1/projects/{project_id}/creative-workspaces/commerce-preroll:
    get:
      operationId: getLatestCommercePrerollWorkspace
      summary: Restore the latest durable commerce pre-roll workspace in the Project.
      parameters: [{$ref: '#/components/parameters/ProjectID'}]
      responses:
        '200':
          description: Task, source Intake, latest Draft and generation Attempt lineage.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommercePrerollWorkspace'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/creative/v1/projects/{project_id}/creative-tasks/{task_id}/commerce-preroll:
    get:
      operationId: getCommercePrerollWorkspace
      summary: Read one commerce pre-roll workspace by its real CreativeTask ID.
      parameters:
        - {$ref: '#/components/parameters/ProjectID'}
        - {$ref: '#/components/parameters/TaskID'}
      responses:
        '200':
          description: Commerce pre-roll workspace.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommercePrerollWorkspace'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
...
  securitySchemes:
    cookiesAuth: {type: http, scheme: bearer}
  parameters:
...
    ProjectID: {name: project_id, in: path, required: true, schema: {type: string, minLength: 1}}
...
    TaskID: {name: task_id, in: path, required: true, schema: {type: string, minLength: 1}}
...
  schemas:
...
    AssetVersionRef:
      type: object
      additionalProperties: false
      required: [asset_id, version]
      properties:
        asset_id: {type: string, minLength: 1}
        version: {type: integer, minimum: 1}
...
    CommercePrerollWorkspace:
      type: object
      additionalProperties: true
      required:
        - task
        - intake
        - video_draft
        - commerce_preroll_generation_attempts
      properties:
        task: {type: object}
        intake: {type: object}
        video_draft:
          type: object
          required: [contract_version, task_id, revision, commerce_preroll]
          properties:
            contract_version: {const: creative-video-draft/v1}
            task_id: {type: string}
            revision: {type: integer, minimum: 1}
            commerce_preroll:
              type: object
              required: [contract_version, task_id, revision, input_snapshot, input_hash, plan, readiness]
              properties:
                contract_version: {const: creative-commerce-preroll-draft/v1}
                task_id: {type: string}
                revision: {type: integer, minimum: 1}
                input_snapshot: {type: object}
                input_hash: {$ref: '#/components/schemas/SHA256Ref'}
                plan: {type: object}
                approval: {$ref: '#/components/schemas/VideoGenerationApproval'}
                readiness: {type: object}
        commerce_preroll_generation_attempts:
          type: array
          items:
            type: object
            required: [id, task_id, draft_revision, template_ref, prompt_hash, generation_spec_hash, provider_job_id, created_at]
            properties:
              id: {type: string}
              task_id: {type: string}
              draft_revision: {type: integer, minimum: 1}
              template_ref: {$ref: '#/components/schemas/CreativeTemplateRef'}
              prompt_hash: {$ref: '#/components/schemas/SHA256Ref'}
              generation_spec_hash: {$ref: '#/components/schemas/SHA256Ref'}
              provider_job_id: {type: string}
              retry_of_attempt_id: {type: string}
              output_asset_version: {$ref: '#/components/schemas/AssetVersionRef'}
              created_at: {type: string, format: date-time}
...
    CreativeTemplateRef:
      type: object
      additionalProperties: false
      required: [template_id, template_version]
      properties:
        template_id:
          enum:
            - commerce.product-cut
            - commerce.window-reveal
            - commerce.one-click
            - commerce.miniature
            - commerce.device-summon
        template_version: {type: integer, const: 1}
...
    VideoGenerationApproval:
      type: object
      additionalProperties: false
      required: [contract_version, task_id, generation_spec_hash, confirmed_items, confirmed_by, confirmed_at]
      properties:
        contract_version: {const: creative-generation-approval/v1}
        task_id: {type: string, minLength: 1}
        generation_spec_hash: {$ref: '#/components/schemas/SHA256Ref'}
        confirmed_items:
          type: array
          minItems: 6
          maxItems: 6
          uniqueItems: true
          items: {enum: [product, template, motion, result, prompt, paid_generation]}
        confirmed_by: {type: string, minLength: 1}
        confirmed_at: {type: string, format: date-time}
    SHA256Ref:
      type: string
      pattern: '^sha256:[a-f0-9]{64}$'
...
  responses:
...
    NotFound: {description: The scoped resource does not exist.}
...